can add and delete from favorite

+++ backend/api/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
+
 from rest_framework import mixins, permissions, viewsets, status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.filters import SearchFilter
-from rest_framework.permissions import AllowAny, IsAuthenticatedOrReadOnly
+from rest_framework.permissions import (
+    AllowAny, IsAuthenticatedOrReadOnly, IsAuthenticated
+)
 
 from api.serializers import (
     RecipeSerializer, RecipeCreateSerializer,
@@ -33,7 +36,7 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
 
 
 class RecipeViewSet(viewsets.ModelViewSet):
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,
+    permission_classes = (IsAuthenticatedOrReadOnly,
                           AuthorOrReadOnly,)
     pagination_class = ResultsSetPagination
 
@@ -58,47 +61,60 @@ class RecipeViewSet(viewsets.ModelViewSet):
         )
 
     @staticmethod
-    def add_to(self, Model, message, **kwargs):
-        recipe = get_object_or_404(Recipe, pk=kwargs['pk'])
+    def add_to(self, Model, pk):
+        if Model.objects.filter(
+            user=self.request.user,
+            recipe=pk
+        ).exists():
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        recipe = get_object_or_404(Recipe, pk=pk)
         Model.objects.create(
                 user=self.request.user,
                 recipe=recipe
             )
 
         return Response(
-                {'detail': message},
                 ShortRecipeSerializer(recipe).data,
                 status=status.HTTP_201_CREATED
             )
 
     @staticmethod
-    def delete_from(self, Model, message, **kwargs):
-        recipe = get_object_or_404(Recipe, pk=kwargs['pk'])
+    def delete_from(self, Model, pk):
+        if not Model.objects.filter(
+            user=self.request.user,
+            recipe=pk
+        ).exists():
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        recipe = get_object_or_404(Recipe, pk=pk)
         Model.objects.get(
                 user=self.request.user,
                 recipe=recipe
             ).delete()
 
         return Response(
-            {'detail': message},
             status=status.HTTP_204_NO_CONTENT
         )
 
     @action(
-        detail=True, methods=('post', 'delete',)
+        detail=True, methods=('post', 'delete',),
+        permission_classes=(IsAuthenticated,)
     )
-    def favorite(self, request, **kwargs):
+    def favorite(self, request, pk=None):
 
         if request.method == 'POST':
-            self.add_to(
+            return self.add_to(
                 self, Model=Favorite,
-                message='Рецепт теперь в избранном'
+                pk=pk
             )
 
         if request.method == 'DELETE':
-            self.delete_from(
+            return self.delete_from(
                 self, Model=Favorite,
-                message='Рецепт удалён из избранного'
+                pk=pk
             )
 '''
     @action(
@@ -106,15 +122,13 @@ class RecipeViewSet(viewsets.ModelViewSet):
     )
     def shopping_cart(self, request, pk):
         if request.method == 'POST':
-            self.add_to(
+            return self.add_to(
                 self, Model=Cart,
-                message='Рецепт теперь в корзине'
             )
 
         if request.method == 'DELETE':
-            self.delete_from(
+            return self.delete_from(
                 self, Model=Cart,
-                message='Рецепт удалён из корзины'
             )
 
     @action(
